Move sync defines out from storage tree.
